feature-AddNewEconomyEnginePt1 (#5)
* Pt1

* Added txt data

* Faze2

* Faze3

* Faze4

* Fix DivideByZero warnings

* Refactoring Sim Engine & Validation

* Add Warehouse API Controll

+++ RPGEconomy.API/Controllers/ProductTypesController.cs
@@ -61,5 +61,5 @@ public class ProductTypesController : ControllerBase
 public record CreateProductTypeRequest(
     string Name,
     string Description,
-    double BasePrice,
+    decimal BasePrice,
     double WeightPerUnit);
